lp-698 : reorganise packages
