+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/common/Nationality.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.partnership;
+package uk.gov.companieshouse.limitedpartnershipsapi.model.common;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
