+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/generalpartner/dto/GeneralPartnerDataDto.java
@@ -6,10 +6,10 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Past;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.Country;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.Nationality;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.dto.AddressDto;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.ValidEnum;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.common.Country;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.common.Nationality;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.common.dto.AddressDto;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.EnumValid;
 
 import java.time.LocalDate;
 
@@ -63,11 +63,11 @@ public class GeneralPartnerDataDto {
     private LocalDate dateOfBirth;
 
     @JsonProperty(NATIONALITY1_FIELD)
-    @ValidEnum(message = "First nationality must be valid")
+    @EnumValid(message = "First nationality must be valid")
     private Nationality nationality1;
 
     @JsonProperty(NATIONALITY2_FIELD)
-    @ValidEnum(message = "Second nationality must be valid")
+    @EnumValid(message = "Second nationality must be valid")
     private Nationality nationality2;
 
     @JsonProperty(NOT_DISQUALIFIED_STATEMENT_CHECKED_FIELD)
@@ -99,7 +99,7 @@ public class GeneralPartnerDataDto {
     private String legalEntityRegisterName;
 
     @JsonProperty(LEGAL_ENTITY_REGISTRATION_LOCATION_FIELD)
-    @ValidEnum(message = "Legal entity registration location must be valid")
+    @EnumValid(message = "Legal entity registration location must be valid")
     private Country legalEntityRegistrationLocation;
 
     @JsonProperty(LEGAL_FORM_FIELD)
