+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/incorporation/IncorporationKind.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.partnership;
+package uk.gov.companieshouse.limitedpartnershipsapi.model.incorporation;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
