+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/incorporation/dto/IncorporationDataDto.java
@@ -1,10 +1,10 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.dto;
+package uk.gov.companieshouse.limitedpartnershipsapi.model.incorporation.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.NotNull;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.IncorporationKind;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.ValidEnum;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.incorporation.IncorporationKind;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.EnumValid;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
@@ -12,7 +12,7 @@ public class IncorporationDataDto {
 
     @JsonInclude(NON_NULL)
     @JsonProperty("kind")
-    @ValidEnum(message = "Kind must be valid")
+    @EnumValid(message = "Kind must be valid")
     @NotNull(message = "Kind must not be null")
     private IncorporationKind kind;
