+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/incorporation/dto/IncorporationDto.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.dto;
+package uk.gov.companieshouse.limitedpartnershipsapi.model.incorporation.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
