+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/partnership/dao/DataDao.java
@@ -1,6 +1,7 @@
 package uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.dao;
 
 import org.springframework.data.mongodb.core.mapping.Field;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.common.dao.AddressDao;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.PartnershipType;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.Term;
