+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/partnership/dto/LimitedPartnershipPatchDto.java
@@ -5,12 +5,13 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.common.dto.AddressDto;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.Jurisdiction;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.PartnershipNameEnding;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.PartnershipType;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.Term;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.EnumValid;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.NameSize;
-import uk.gov.companieshouse.limitedpartnershipsapi.model.validator.ValidEnum;
 
 import java.util.List;
 
@@ -31,7 +32,7 @@ public class LimitedPartnershipPatchDto {
     private String partnershipName;
 
     @JsonProperty("name_ending")
-    @ValidEnum(message = "Name ending must be valid")
+    @EnumValid(message = "Name ending must be valid")
     private PartnershipNameEnding nameEnding;
 
     @JsonProperty("email")
@@ -39,11 +40,11 @@ public class LimitedPartnershipPatchDto {
     private String email;
 
     @JsonProperty("partnership_type")
-    @ValidEnum(message = "Partnership type must be valid")
+    @EnumValid(message = "Partnership type must be valid")
     private PartnershipType partnershipType;
 
     @JsonProperty("jurisdiction")
-    @ValidEnum(message = "Jurisdiction must be valid")
+    @EnumValid(message = "Jurisdiction must be valid")
     private Jurisdiction jurisdiction;
 
     @JsonProperty("registered_office_address")
@@ -51,7 +52,7 @@ public class LimitedPartnershipPatchDto {
     private AddressDto registeredOfficeAddress;
 
     @JsonProperty("term")
-    @ValidEnum(message = "Term must be valid")
+    @EnumValid(message = "Term must be valid")
     private Term term;
 
     @JsonProperty("principal_place_of_business_address")
